requirements.txt
colorama
discord.py
pystyle
requests
tqdm
random